services/local_file_organizer/app/requirements.txt
watchdog
PyYAML
loguru